Display the CloudFormation outputs in the `deployment` command
The `bref deployment` command currently displays the events of the latest deployments.

With this change the command will also display the "outputs" of the CloudFormation stack.

The practical side of this is that we can now get the URL of a HTTP application that we just deployed. Without this we need to log into the AWS console and check the stack in the CloudFormation dashboard… Less practical.