Updated FPM php to cater for api gateways authorisers
This request context is needed if a user is using the authentication type of IAM_ROLE, when a user auths through Api Gateway it will then authorise the user and pass their information (securely as request context). At the moment the file misses this so a php function cant use the authorised details correctly.