Revise Lambda/Request context within PHP-FPM handler
Currently we have access to the Request context within the PHP-FPM handler request thanks to the inclusion of the `LAMBDA_CONTEXT` env. var present per request.
This is great but sadly does not provide exposure to the true Lambda Context which is supplied per Lambda invocation, only the request context present in the event.

I feel that this naming of this env. var is a little misleading and as such this PR deprecates the current usage of this `LAMBDA_CONTEXT` env. var.
In it’s place it provides two new env. vars. `LAMBDA_INVOCATION_CONTEXT` and `LAMBDA_REQUEST_CONTEXT` which provide access to the true Lambda Context and Event Request Context.